chore: moved setup of environment into setup package.
